Agents/SummarizationAgent/requirements.txt
google-generativeai
PyPDF2
python-docx
easyocr